MAINT: move several utils to _util from _helpers
